chore: add ad
