+++ packages/ad/package.json
@@ -45,9 +45,8 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/addon-actions": "3.4.1",
-    "@storybook/react-native": "3.4.1",
     "@times-components/eslint-config-thetimes": "0.4.1",
+    "@times-components/storybook": "0.9.3",
     "babel-core": "6.26.0",
     "babel-plugin-add-react-displayname": "0.0.5",
     "babel-plugin-flow-react-proptypes": "22.0.0",
