+++ packages/icons/icons.showcase.js
@@ -39,18 +39,20 @@ const renderIcon = color => args => {
   const name = args[0];
   const Icon = args[1];
 
-  return <View style={styles.icon} key={name}>
-    <Icon
-      width={50}
-      height={50}
-      fillColour={color(`Icon ${name} fill`, Icon.defaultProps.fillColour)}
-      strokeColour={color(
-        `Icon ${name} strike`,
-        Icon.defaultProps.strokeColour
-      )}
-    />
-    <Text style={styles.label}>{name}</Text>
-  </View>;
+  return (
+    <View style={styles.icon} key={name}>
+      <Icon
+        width={50}
+        height={50}
+        fillColour={color(`Icon ${name} fill`, Icon.defaultProps.fillColour)}
+        strokeColour={color(
+          `Icon ${name} strike`,
+          Icon.defaultProps.strokeColour
+        )}
+      />
+      <Text style={styles.label}>{name}</Text>
+    </View>
+  );
 };
 
 export default {
